refactor: 중복 체크 함수 분리
